# Conflicts:
#	server/src/main/java/org/elasticsearch/cluster/routing/allocation/decider/NodeReplacementAllocationDecider.java

+++ server/src/test/java/org/elasticsearch/action/search/SearchPhaseControllerTests.java
@@ -489,7 +489,7 @@ public class SearchPhaseControllerTests extends ESTestCase {
         for (SearchPhaseResult shardResult : results.asList()) {
             TopDocs topDocs = shardResult.queryResult().topDocs().topDocs;
             assert topDocs.totalHits.relation == Relation.EQUAL_TO;
-            resultCount += topDocs.totalHits.value;
+            resultCount += (int) topDocs.totalHits.value;
         }
         return resultCount;
     }
